Move RootSpanFields to api
